+++ Battleship/ButtonTexture.h
@@ -16,7 +16,7 @@ public:
 	// Create a texture button with no text
 	ButtonTexture(const sf::Vector2f& pos, const sf::Vector2f& size, const std::string& text, sf::Texture* normal_tex, sf::Texture* press_tex, unsigned state);
 	// Updates the button state
-	void update(const sf::Vector2i& mouse_pos, sf::RenderWindow& window);
+	void update(const sf::Vector2f& mouse_pos, sf::RenderWindow& window);
 	// Changes the size of the button to let the user know they're hovering over the button
 	void hover(unsigned x_off, unsigned y_off, unsigned char_off);
 	// Change the texture/size of the button to let the user know they've pressed the button
