+++ Battleship/TableScrollable.h
@@ -26,7 +26,7 @@ public:
 	// Creates a table with three columns for keeping statistics on the game
 	TableScrollable(sf::Vector2f pos, sf::Vector2f size, unsigned row_spacing, sf::Font& font, std::vector<sf::Texture*> textures, int rows);
 	// draw & take in mouse_pos for checking embeded buttons
-	void draw(sf::RenderWindow& win, sf::Vector2i& mouse_pos);
+	void draw(sf::RenderWindow& win, sf::Vector2f& mouse_pos);
 	// add an entry to the Hits table
 	void add_entry(sf::Vector2i player_hit, sf::Vector2i enemy_hit, bool& successful_player_attack, bool& successful_enemy_attack);
 	void process_click(sf::Vector2f& mouse_pos);
